pyproject.toml
[tool.poetry]
name = "supplier-invoice-automation"
version = "0.1.0"
description = "Automated invoice extraction service"
authors = ["Development Team"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11.9"
fastapi = "^0.116.1"
uvicorn = "^0.35.0"
python-dotenv = "^1.1.1"
loguru = "^0.7.3"
pydantic-settings = "^2.7.0"
pytest = "^8.3.5"
pytest-asyncio = "^0.24.0"
fakeredis = "^2.25.1"
httpx = "^0.28.0"
surya-ocr = "^0.6.9"
python-multipart = "^0.0.17"
pillow = "^10.4.0"
aiofiles = "^24.1.0"
pymupdf = "^1.24.12"
redis = "^5.0.7"
tenacity = "^8.3.0"
google-generativeai = "^0.7.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
